mysite03 #6-board

+++ mysite03/src/main/java/com/poscoict/mysite/controller/BoardController.java
@@ -1,6 +1,8 @@
-package com.douzone.mysite.controller;
+package com.poscoict.mysite.controller;
 
 import java.util.Map;
+
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -25,8 +27,10 @@ public class BoardController {
 	@RequestMapping("")
 	public String index(@RequestParam(value = "p", required = true, defaultValue = "1") Integer page, @RequestParam(value = "kwd", required = true, defaultValue = "") String keyword, Model model) {
 		Map<String, Object> map = boardService.getContentsList(page, keyword);
+		
 		model.addAttribute("map", map);
-		// model.addAllAttributes(map);
+		// 동일: model.addAllAttributes(map);
+		
 		return "board/index";
 	}
 
@@ -34,46 +38,84 @@ public class BoardController {
 	public String view(@PathVariable("no") Long no, Model model) {
 		BoardVo boardVo = boardService.getContents(no);
 		model.addAttribute("boardVo", boardVo);
+		
 		return "board/view";
 	}
 
 	@RequestMapping("/delete/{no}")
-	public String delete(@AuthUser UserVo authUser, @PathVariable("no") Long boardNo, @RequestParam(value = "p", required = true, defaultValue = "1") Integer page, @RequestParam(value = "kwd", required = true, defaultValue = "") String keyword) {
+	public String delete(HttpSession session, @PathVariable("no") Long boardNo, @RequestParam(value = "p", required = true, defaultValue = "1") Integer page, @RequestParam(value = "kwd", required = true, defaultValue = "") String keyword) {
+		/* access controller */
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/";
+		}
+		
 		boardService.deleteContents(boardNo, authUser.getNo());
+
 		return "redirect:/board?p=" + page + "&kwd=" + WebUtil.encodeURL(keyword, "UTF-8");
 	}
 
 	@RequestMapping(value = "/update/{no}")
-	public String modify(@AuthUser UserVo authUser, @PathVariable("no") Long no, Model model) {
+	public String update(HttpSession session, @PathVariable("no") Long no, Model model) {
+		/* access controller */
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/";
+		}
+		
 		BoardVo boardVo = boardService.getContents(no, authUser.getNo());
 		model.addAttribute("boardVo", boardVo);
-		return "board/modify";
+
+		return "board/update";
 	}
 
-	@RequestMapping(value = "/modify", method = RequestMethod.POST)
-	public String modify(@AuthUser UserVo authUser, @ModelAttribute BoardVo boardVo, @RequestParam(value = "p", required = true, defaultValue = "1") Integer page, @RequestParam(value = "kwd", required = true, defaultValue = "") String keyword) {
+	@RequestMapping(value = "/update", method = RequestMethod.POST)
+	public String update(HttpSession session, BoardVo boardVo, @RequestParam(value = "p", required = true, defaultValue = "1") Integer page, @RequestParam(value = "kwd", required = true, defaultValue = "") String keyword) {
+		/* access controller */
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/";
+		}
+		
 		boardVo.setUserNo(authUser.getNo());
-		boardService.modifyContents(boardVo);
+		boardService.updateContents(boardVo);
+
 		return "redirect:/board/view/" + boardVo.getNo() + "?p=" + page + "&kwd=" + WebUtil.encodeURL(keyword, "UTF-8");
 	}
 
-	@Auth
 	@RequestMapping(value = "/write", method = RequestMethod.GET)
-	public String write() {
+	public String write(HttpSession session) {
+		/* access controller */
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/";
+		}
+		
 		return "board/write";
 	}
 
-	@Auth
 	@RequestMapping(value = "/write", method = RequestMethod.POST)
-	public String write(@AuthUser UserVo authUser, @ModelAttribute BoardVo boardVo, @RequestParam(value = "p", required = true, defaultValue = "1") Integer page, @RequestParam(value = "kwd", required = true, defaultValue = "") String keyword) {
+	public String write(HttpSession session, @ModelAttribute BoardVo boardVo, @RequestParam(value = "p", required = true, defaultValue = "1") Integer page, @RequestParam(value = "kwd", required = true, defaultValue = "") String keyword) {
+		/* access controller */
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/";
+		}
+		
 		boardVo.setUserNo(authUser.getNo());
 		boardService.addContents(boardVo);
+
 		return "redirect:/board?p=" + page + "&kwd=" + WebUtil.encodeURL(keyword, "UTF-8");
 	}
 
-	@Auth
 	@RequestMapping(value = "/reply/{no}")
-	public String reply(@PathVariable("no") Long no, Model model) {
+	public String reply(HttpSession session, @PathVariable("no") Long no, Model model) {
+		/* access controller */
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/";
+		}
+
 		BoardVo boardVo = boardService.getContents(no);
 		boardVo.setOrderNo(boardVo.getOrderNo() + 1);
 		boardVo.setDepth(boardVo.getDepth() + 1);
